main.js に Memo class 以外を移動
